feat: implement InteractionsController and SessionsController for improved API structure and functionality
